Use qint64 instead of size_t for the folder size calculation, otherwise folders >2GB won't show up correctly on 32 Bit systems.
I hope I found all places were the conversion was necessary.
I couldn't really test this, as i don't have a 64 Bit system, but I should
work.

BUG: 151155


svn path=/trunk/KDE/kdepim/; revision=751020

+++ favoritefolderview.cpp
@@ -103,7 +103,7 @@ void FavoriteFolderViewItem::updateCount()
   }
 
   if ( !folder()->noContent() ) {
-    int size = folder()->storage()->folderSize();
+    qint64 size = folder()->storage()->folderSize();
     if ( size != folderSize() ) {
       setFolderSize( size );
       repaint = true;
